implemented remote Matlabs
to start an instance of Matlab on a remote machine, supply an IP address
for the `address` parameter of `Matlab.__init__`. The remote machine
must be reachable by SSH. Supply a user account in the `user`
parameter (if necessary).

Transplant will open an SSH connection and start Matlab on the remote
machine, then connect to that Matlab using a ZMQ/TCP connection.